requirements.txt
Click
python-frontmatter